Change, use dynamic_cast to enter node
- localizer -- Fix bug
- README -- Add note of `require` function

The usage of a nested "visitor" struct function within another struct is not correct because it skips over nodes that we are not interested in within the overridden struct, which is the default value of the base class.

The fix for this problem is to use dynamic_cast to check if the next node is the one of interest, instead of calling the "visitor" function again.