Fix: Alias not found when the required Luau file is not present in the same path as .luaurc
Search for the config file by traversing directories backward up to the root (where the main file is located) to gather the configurations. The root of .luaurc is considered the directory of the main file, not the current working directory, similar to how Luau runs. This is because it is easier to manage. In the future, there might be an option to customize this behavior.