Fix: When main file does not require a module, the output is empty
This issue occurs because the file aggregator lists items using a topological sort. However, the sort returns an empty vector due to the absence of added dependencies, causing the aggregator not to generate the main function for execution in the output. As a result, only a bare call to the main function remains.

The solution is to check whether there is only one module. If so, manually create a single aggregation order instead of calling `load_order`.